backend-service/requirements.txt
alembic==0.8.7
amqp==1.4.9
aniso8601==1.1.0
anyjson==0.3.3
apipkg==1.4
astroid==1.4.8
backports.shutil-get-terminal-size==1.0.0
billiard==3.3.0.23
blinker==1.4
celery==3.1.23
click==6.6
colorama==0.3.7
colorlog==2.7.0
cov-core==1.15.0
coverage==4.2
decorator==4.0.10
eventlet==0.19.0
execnet==1.4.1
factory-boy==2.7.0
fake-factory==0.6.0
flake8==3.0.4
Flask==0.11.1
Flask-Cors==2.1.3
Flask-Migrate==2.0.0
Flask-RESTful==0.3.5
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-Testing==0.5.0
gevent==1.1.2
greenlet==0.4.10
gunicorn==19.6.0
ipython==5.1.0
ipython-genutils==0.1.0
isort==4.2.5
itsdangerous==0.24
Jinja2==2.8
kombu==3.0.35
lazy-object-proxy==1.2.2
logutils==0.3.3
Mako==1.0.4
MarkupSafe==0.23
marshmallow==2.9.1
mccabe==0.5.2
newrelic==2.68.0.50
nose2==0.6.5
path.py==8.2.1
pep8==1.7.0
pexpect==4.2.0
pickleshare==0.7.4
prompt-toolkit==1.0.6
psycopg2==2.6.2
ptyprocess==0.5.1
py==1.4.31
pycodestyle==2.0.0
pycrypto==2.6.1
pyflakes==1.2.3
Pygments==2.1.3
PyJWT==1.4.2
pylint==1.6.4
pytest==2.9.2
pytest-cache==1.0
pytest-pep8==1.0.6
pytest-pythonpath==0.7
python-dateutil==2.5.3
python-editor==1.0.1
pytz==2016.6.1
requests==2.11.1
simplegeneric==0.8.1
simplejson==3.8.2
six==1.10.0
SQLAlchemy==1.0.14
SQLAlchemy-Continuum==1.2.4
SQLAlchemy-Utils==0.32.9
suds-jurko==0.6
traitlets==4.2.2
vine==1.1.1
wcwidth==0.1.7
webargs==1.3.4
Werkzeug==0.11.10
wrapt==1.10.8
